refactor: migrate business logic from routes to controllers
- Moved business logic from routes to dedicated controller files for better code organization and separation of concerns.
